requirements.txt
streamlit
google-generativeai
ollama
rich
asyncio
python-dotenv
